_deprecated/livekit-agent/requirements.txt
# LiveKit Agent Dependencies - Production Ready
# Python 3.11+ required
# All versions pinned per project policy (CLAUDE.md)

# Core LiveKit libraries
livekit==1.0.16
livekit-agents==1.3.5
livekit-api==1.0.7
livekit-protocol==1.0.8

# LiveKit plugins
livekit-plugins-deepgram==1.3.5
livekit-plugins-cartesia==1.3.5
livekit-plugins-openai==1.3.5
livekit-plugins-anthropic==1.3.5
livekit-plugins-google==1.3.5
livekit-plugins-silero==1.3.5

# Vision & Image Processing (perceptual hashing + frame encoding)
opencv-python==4.12.0.88
Pillow==11.1.0
imagehash==4.3.2
numpy==2.2.6

# HTTP Client & JWT (backend communication)
httpx==0.28.1
h2==4.3.0
pyjwt==2.10.1

# Configuration & Environment
python-dotenv==1.1.1
pydantic==2.12.0
pydantic-settings==2.11.0

# Database & Cache (knowledge base + rate limiting)
asyncpg==0.30.0
redis==6.4.0

# AI Providers (direct SDK access)
openai==2.3.0
anthropic==0.69.0
google-generativeai==0.8.5

# Logging & Monitoring
structlog==25.4.0

# Testing
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-timeout==2.4.0
